dont work multi def
