+++ Enclave/include/ycsb.h
@@ -208,7 +208,7 @@ RETRY:
         // ycsb_tuple.id_ = i;
         Tuple* tmp = new Tuple(key.view(), 0);
         tmp->init();
-        Table.insert_value(tmp);
+        Table.put(key.view(),tmp,1);
         // Masstrees[get_storage(Storage::YCSB)].insert_value(key.view(), tmp);
       }
     }
